Robustness added to PUT method. NEEDS REFACTORING. Ref #12

+++ app/ruby/views.py
@@ -129,11 +129,23 @@ def update_ruby_challenge(id):
     update_data = json.loads(request.form.get('challenge'))['challenge']
     old_challenge = get_challenge(id).get_dict()
     source_code_name = f"{update_data['source_code_file_name']}.rb"
-    source_code_path_tmp = f"public/{source_code_name}"
     source_test_name = f"{update_data['test_suite_file_name']}.rb"
+    source_code_path_destiny = f"public/challenge/{source_code_name}"
+    source_test_path_destiny = f"public/challenge/{source_test_name}"
+    
+    # Check if file name already exists
+    if (file_exists(source_code_path_destiny) and old_challenge['code'] != source_code_path_destiny): 
+        return make_response(jsonify({'code': 'code name already exists'}), 400)
+    if (file_exists(source_test_path_destiny) and old_challenge['tests_code'] != source_test_path_destiny):
+        return make_response(jsonify({'test': 'test name already exists'}), 400)
+    del source_code_path_destiny, source_test_path_destiny
+    
+    source_code_path_tmp = f"public/{source_code_name}"
     source_test_path_tmp = f"public/{source_test_name}"
+
     delete_keys(update_data, ['source_code_file_name', 'test_suite_file_name'])
 
+    # If there is a new code file, check if it compiles
     if 'source_code_file' in request.files:
         request.files['source_code_file'].save(dst=source_code_path_tmp)
         if os.path.isfile(source_code_path_tmp) and not compiles(source_code_path_tmp):
@@ -142,6 +154,7 @@ def update_ruby_challenge(id):
     else:
         update_file_name(old_challenge, 'code', source_code_name, update_data)
 
+    # If there is a new test suite, check dependencies and if it fails.
     if 'test_suite_file' in request.files:
         request.files['test_suite_file'].save(dst=source_test_path_tmp)
         if os.path.isfile(source_test_path_tmp) and not tests_fail(source_test_path_tmp):
@@ -152,17 +165,26 @@ def update_ruby_challenge(id):
             return make_response(jsonify({'tests': "dependencies failed to compile"}), 400)
     else:
         if not dependencies_ok(old_challenge['tests_code'], os.path.basename(source_code_name.split('.')[0])):
+            os.remove(source_test_path_tmp)
             return make_response(jsonify({'tests': "dependencies failed to compile"}), 400)
         update_file_name(old_challenge, 'tests_code', source_test_name, update_data)
         
+    #Update DB info (complexity, objetive, etc) and remove tmp data in failure case
     if update_challenge(id, update_data) < 1:
+        if os.path.isfile(source_test_path_tmp):
+            os.remove(source_test_path_tmp)
+        if os.path.isfile(source_code_path_tmp):
+            os.remove(source_code_path_tmp)
         return make_response(jsonify({'challenge': "update failed check input data"}), 400)
 
+    # Update test and code files if they exists
     update_file(old_challenge, 'code', source_code_path_tmp, source_code_name, update_data)
     update_file(old_challenge, 'tests_code', source_test_path_tmp, source_test_name, update_data)
 
+    # Update challenge with new paths
     update_challenge(id, update_data)
     
+    # Retur updated challenge
     updated_challenge = get_challenge(id).get_dict()
     delete_keys(updated_challenge, ['id'])
     return jsonify({'challenge': updated_challenge})
@@ -188,6 +210,8 @@ def create_challenge(challenge):
     db.session.commit()
 
 def update_challenge(id, changes):
+    if len(changes) == 0:
+        return 1
     result = db.session.query(RubyChallenge).filter_by(id=id).update(changes)
     db.session.commit()
     return result
